Ot193 45 read news

+++ src/main/java/com/alkemy/somosmas/services/impl/NewsServiceImpl.java
@@ -1,5 +1,6 @@
 package com.alkemy.somosmas.services.impl;
 
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -36,7 +37,19 @@ public class NewsServiceImpl implements NewsService {
 
 	@Override
 	public void delete(Long id) {
+		if (!this.newsRepository.existsById(id)) {
+			throw new RuntimeException("Invalid ID");
+		}
 		this.newsRepository.deleteById(id);
 	}
 
+	@Override
+	public NewsDTO getDetailsById(Long id) {
+		Optional<News> newsEntity = this.newsRepository.findById(id); 
+		if (!newsEntity.isPresent()) {
+			throw new RuntimeException("Invalid ID");
+		}
+		NewsDTO result = this.newsMapper.newsEntity2DTO(newsEntity.get());
+		return result;
+	}
 }
